refactor: move client logging endpoints to blueprint
