+++ src/web/app.py
@@ -39,26 +39,6 @@ trading_logger = page_logger.logger
 app = Flask(__name__)
 # Register route blueprints
 register_routes(app)
-# --- Client Error Logging Endpoint ---
-@app.route("/api/log_client_error", methods=["POST"])
-def log_client_error():
-    """Log client-side JS errors from the frontend."""
-    try:
-        data = request.get_json(force=True)
-        page = data.get("page", "unknown")
-        error = data.get("error", "No error message")
-        stack = data.get("stack", "No stack trace")
-        timestamp = data.get("timestamp", datetime.now().isoformat())
-        log_message = f"[CLIENT ERROR] Page: {page} | Error: {error} | Stack: {stack} | Timestamp: {timestamp}"
-        trading_logger.error_logger.error(log_message)
-        log_exception(f"Client error on {page}", error)
-        return create_api_response(message="Error logged successfully")
-    except Exception as e:
-        return handle_api_error(e, "log_client_error endpoint")
-@app.route("/api/frontend_logs", methods=["POST"])
-def frontend_logs():
-    """Alternative endpoint for frontend logging (compatibility)"""
-    return log_client_error()
 # Enable CORS for all routes
 CORS(
     app,
