+++ src/web/routes/__init__.py
@@ -11,6 +11,7 @@ def register_routes(app):
     from .system_routes import system_bp
     from .page_routes import page_bp
     from .telegram_routes import telegram_bp
+    from .logging_routes import logging_bp
     
     # Register blueprints
     app.register_blueprint(analysis_bp)
@@ -18,6 +19,7 @@ def register_routes(app):
     app.register_blueprint(system_bp)
     app.register_blueprint(page_bp)
     app.register_blueprint(telegram_bp)
+    app.register_blueprint(logging_bp)
     
     # Register scalping signals blueprint
     try:
